Move `CryptoProvider` validation into that type
