+++ rustls/src/builder.rs
@@ -1,12 +1,10 @@
 use alloc::format;
-use alloc::vec::Vec;
 use core::fmt;
 use core::marker::PhantomData;
 
 use crate::client::EchMode;
 use crate::crypto::CryptoProvider;
 use crate::error::Error;
-use crate::msgs::handshake::ALL_KEY_EXCHANGE_ALGORITHMS;
 use crate::sync::Arc;
 use crate::time_provider::TimeProvider;
 use crate::versions;
@@ -221,38 +219,7 @@ impl<S: ConfigSide> ConfigBuilder<S, WantsVersions> {
             return Err(Error::General("no usable cipher suites configured".into()));
         }
 
-        if self.provider.kx_groups.is_empty() {
-            return Err(Error::General("no kx groups configured".into()));
-        }
-
-        // verifying cipher suites have matching kx groups
-        let mut supported_kx_algos = Vec::with_capacity(ALL_KEY_EXCHANGE_ALGORITHMS.len());
-        for group in self.provider.kx_groups.iter() {
-            let kx = group.name().key_exchange_algorithm();
-            if !supported_kx_algos.contains(&kx) {
-                supported_kx_algos.push(kx);
-            }
-            // Small optimization. We don't need to go over other key exchange groups
-            // if we already cover all supported key exchange algorithms
-            if supported_kx_algos.len() == ALL_KEY_EXCHANGE_ALGORITHMS.len() {
-                break;
-            }
-        }
-
-        for cs in self.provider.cipher_suites.iter() {
-            let cs_kx = cs.key_exchange_algorithms();
-            if cs_kx
-                .iter()
-                .any(|kx| supported_kx_algos.contains(kx))
-            {
-                continue;
-            }
-            let suite_name = cs.common().suite;
-            return Err(Error::General(alloc::format!(
-                "Ciphersuite {suite_name:?} requires {cs_kx:?} key exchange, but no {cs_kx:?}-compatible \
-                key exchange groups were present in `CryptoProvider`'s `kx_groups` field",
-            )));
-        }
+        self.provider.consistency_check()?;
 
         Ok(ConfigBuilder {
             state: WantsVerifier {
